Remaking whole structure of the project
Due to bad structure, I started spliting the code into more files. There are still some bugs in this code / for example I cant open app if last saved file is GitHub
